+++ lib/livebook/zta/basic_auth.ex
@@ -28,7 +28,8 @@ defmodule Livebook.ZTA.BasicAuth do
   end
 
   @impl true
-  def logout(_name, _socket) do
-    :error
-  end
+  def logout(_name, _socket), do: raise("not implemented")
+
+  @impl true
+  def logout_supported?, do: false
 end
