+++ lib/livebook/zta/pass_through.ex
@@ -12,7 +12,8 @@ defmodule Livebook.ZTA.PassThrough do
   end
 
   @impl true
-  def logout(_name, _socket) do
-    :error
-  end
+  def logout(_name, _socket), do: raise("not implemented")
+
+  @impl true
+  def logout_supported?, do: false
 end
