+++ lib/livebook/zta/tailscale.ex
@@ -30,9 +30,10 @@ defmodule Livebook.ZTA.Tailscale do
   end
 
   @impl true
-  def logout(_name, _socket) do
-    :error
-  end
+  def logout(_name, _socket), do: raise("not implemented")
+
+  @impl true
+  def logout_supported?, do: false
 
   defp authenticate_ip(remote_ip, address) do
     {url, options} =
