--- lib/livebook_web/controllers/logout_html.ex
@@ -1,5 +1,0 @@
-defmodule LivebookWeb.LogoutHTML do
-  use LivebookWeb, :html
-
-  embed_templates "logout_html/*"
-end
